move all viz tool into pyoviz, thus delete the 'obj' directory
