+++ env.yml
@@ -9,9 +9,7 @@ dependencies:
 - numpy
 - pandas
 - matplotlib
-- vtk
 - scipy
-- pyqt
 - boost
 - rbdl
 - dlib
