+++ pyomeca/frame_dependent.py
