+++ setup.py
@@ -3,17 +3,6 @@ from setuptools import setup
 # to install dependencies in a clean conda env, run: `conda env create -f env.yml`
 
 requirements = [
-    'numpy',
-    'pandas',
-    'matplotlib',
-    'vtk',
-    'scipy',
-    'pyqt',
-    'boost',
-    'rbdl',
-    'dlib',
-    'biorbd',
-    'ezc3d'
 ]
 
 setup(
@@ -23,7 +12,7 @@ setup(
     author="Romain Martinez & Pariterre",
     author_email='',
     url='https://github.com/pyomeca/pyomeca',
-    license='MIT license',
+    license='Apache 2.0',
     packages=['pyomeca'],
     install_requires=requirements,
     keywords='pyomeca',
@@ -31,3 +20,5 @@ setup(
         'Programming Language :: Python :: 3.6',
     ]
 )
+
+# TODO: remember to add/update your deps in setup.py
